ver 2.0
Correctly catching rotation angles in x and y axes of a gyroscope in a mobile phone and transfer these data to the server (Main Body of the game on PC).
But have the problem of gimbal lock. It doesn't matter to this project, because the project won't use these data to do real rotations (I suppose).